a few clean-ups and refactoring
