+++ thingsvision/core/extraction/__init__.py
@@ -1,2 +1,1 @@
-from .extractor import get_extractor, get_extractor_from_model
-from .helpers import center_features, normalize_features
+from .helpers import create_custom_extractor, create_model_extractor, get_extractor, get_extractor_from_model, center_features, normalize_features
